Fix on DeleteAll

+++ src/VitalChoice.Data/Repositories/RepositoryAsync.cs
@@ -266,6 +266,8 @@ namespace VitalChoice.Data.Repositories
         public virtual async Task<bool> DeleteAllAsync(CancellationToken cancellationToken,
             ICollection<TEntity> entitySet)
         {
+            if (!entitySet.Any())
+                return false;
             DbSet.RemoveRange(entitySet);
             await Context.SaveChangesAsync(cancellationToken);
             return true;
